move homepage components out, next to main page.tsx
