+++ app/homePageView.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 import React from 'react';
 import LeftNav from '@/app/components/leftNav';
 import Notifications from '@/app/components/notifications';
-import { EntryBanners } from './components/entryBanners';
 import { MicroLayout } from './components/microLayout';
+import { EntryBanners } from './entryBanners';
 
 function HomePageView() {
   return (
