+++ app/page.tsx
@@ -1,4 +1,4 @@
-import { HomePageView } from './views/quickstart-home';
+import { HomePageView } from './homePageView';
 
 export default function HomePage() {
   return <HomePageView />;
